requirements.txt
autopep8==1.5.5
Django ==3.1.1
django-pandas==0.6.2
flake8==3.8.4
psycopg2-binary==2.8.6
pybitflyer==0.1.9
pycparser==2.20
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
websocket-client==0.57.0
plotly==4.14.3
